add tags field to events
