+++ knightassist_mobile_app/lib/src/features/authentication/data/auth_repository.dart
@@ -183,6 +183,15 @@ class AuthRepository {
               backgroundUrl: user['backgroundUrl'] ?? '',
               eventsArray: eventsArray,
               location: user['location'] ?? '',
+                workingHoursPerWeek: user['workingHoursPerWeek'] == null? WorkingHoursPerWeek() : WorkingHoursPerWeek(
+                monday: user['workingHoursPerWeek']['monday'] == null? WeekDay() : WeekDay(start: DateTime.parse(user['workingHoursPerWeek']['monday']['start']), end: DateTime.parse(user['workingHoursPerWeek']['monday']['end'])),
+                tuesday: user['workingHoursPerWeek']['tuesday'] == null? WeekDay() : WeekDay(start: DateTime.parse(user['workingHoursPerWeek']['tuesday']['start']), end: DateTime.parse(user['workingHoursPerWeek']['tuesday']['end'])),
+                wednesday: user['workingHoursPerWeek']['wednesday'] == null? WeekDay() : WeekDay(start: DateTime.parse(user['workingHoursPerWeek']['wednesday']['start']), end: DateTime.parse(user['workingHoursPerWeek']['wednesday']['end'])),
+                thursday: user['workingHoursPerWeek']['thursday'] == null? WeekDay() : WeekDay(start: DateTime.parse(user['workingHoursPerWeek']['thursday']['start']), end: DateTime.parse(user['workingHoursPerWeek']['thursday']['end'])),
+                friday: user['workingHoursPerWeek']['friday'] == null? WeekDay() : WeekDay(start: DateTime.parse(user['workingHoursPerWeek']['friday']['start']), end: DateTime.parse(user['workingHoursPerWeek']['friday']['end'])),
+                saturday: user['workingHoursPerWeek']['saturday'] == null? WeekDay() : WeekDay(start: DateTime.parse(user['workingHoursPerWeek']['saturday']['start']), end: DateTime.parse(user['workingHoursPerWeek']['saturday']['end'])),
+                sunday: user['workingHoursPerWeek']['sunday'] == null? WeekDay() : WeekDay(start: DateTime.parse(user['workingHoursPerWeek']['sunday']['start']), end: DateTime.parse(user['workingHoursPerWeek']['sunday']['end'])),
+              ),
               recoveryTokenForOrg: user['recoveryTokenForORG'],
               confirmTokenForOrg: user['confirmTokenForORG'],
               emailTokenForOrg: user['EmailTokenForORG'],
